Update abx-micro-prescription.sql
Updated Line 108 to  be in line to be inline to Seymour's Suspicion of Infection Time.

"We required the combination of culture and antibiotic start time to occur within a specific time epoch. If the antibiotic was given first, the culture sampling must have been obtained within 24 hours. If the culture sampling was first, the antibiotic must have been ordered within 72 hours. The “onset” of infection was defined as the time at which the first of these 2 events occurred"

Full Text of Seymour et al reference: https://jamanetwork.com/journals/jama/fullarticle/2492875